Rolling Appender
Added a rolling appender function for basic log file management.

Most applications that are concerned with this behavior should consider sending
messages to syslog and use traditional logrotate mechanisms to manage log
files. However, there is a niche of applications that could benefit from this
feature in a compact logging library.